Fixed the problem with lightblue and one last remaining problem
The lightblue problem was caused by the ID got by the previous request was the whole objectid (in JSON format, not only the number) and the Find command executed to see if it will update (if already exist) or insert was not finding the object this way BUT during the insert MongoDB evaluated the index and it got the duplicatekey exection.

The other problem was related to json parse on python with the string split of the framework. Some other problems on tests were pending the other issue fix, solved the last of them (even need to enhance the debug feature of the pyresttest)